# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.AddressType:
    enum:
    - BILLING
    - SHIPPING
    - BOTH
    type: string
    x-enum-varnames:
    - AddressTypeBilling
    - AddressTypeShipping
    - AddressTypeBoth
...
  entity.Customer:
    properties:
      addresses:
        items:
          $ref: '#/definitions/entity.CustomerAddress'
        type: array
      code:
        type: string
      contacts:
        items:
          $ref: '#/definitions/entity.CustomerContact'
        type: array
      created_at:
        type: string
      credit_limit:
        type: number
      current_debt:
        type: number
      email:
        type: string
      id:
        type: integer
      loyalty_points:
        type: integer
      loyalty_tier:
        $ref: '#/definitions/entity.CustomerLoyaltyTier'
      name:
        type: string
      notes:
        type: string
      phone_number:
        type: string
      sales_orders:
        items:
          $ref: '#/definitions/entity.SalesOrder'
        type: array
      tax_id:
        type: string
      type:
        $ref: '#/definitions/entity.CustomerType'
      updated_at:
        type: string
    type: object
  entity.CustomerAddress:
    properties:
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      is_default:
        type: boolean
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
      type:
        $ref: '#/definitions/entity.AddressType'
      updated_at:
        type: string
    type: object
  entity.CustomerContact:
    properties:
      email:
        type: string
      name:
        type: string
      phone_number:
        type: string
      position:
        type: string
    type: object
  entity.CustomerDebt:
    properties:
      last_payment_amount:
        type: number
      last_payment_date:
        type: string
      overdue_debt:
        type: number
      total_debt:
        type: number
      upcoming_payments:
        type: number
    type: object
  entity.CustomerLoyaltyTier:
    enum:
    - STANDARD
    - SILVER
    - GOLD
    - PLATINUM
    type: string
    x-enum-varnames:
    - CustomerLoyaltyTierStandard
    - CustomerLoyaltyTierSilver
    - CustomerLoyaltyTierGold
    - CustomerLoyaltyTierPlatinum
  entity.CustomerOrderHistory:
    properties:
      average_order_value:
        type: number
      first_order_date:
        type: string
      frequent_items:
        items:
          type: string
        type: array
      last_order_date:
        type: string
      total_orders:
        type: integer
      total_spent:
        type: number
    type: object
  entity.CustomerType:
    enum:
    - INDIVIDUAL
    - CORPORATE
    - RESELLER
    - WHOLESALER
    type: string
    x-enum-varnames:
    - CustomerTypeIndividual
    - CustomerTypeCorporate
    - CustomerTypeReseller
    - CustomerTypeWholesaler
...
  entity.SalesOrder:
    properties:
      billing_address:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/entity.User'
      created_by_id:
        type: integer
      customer:
        allOf:
        - $ref: '#/definitions/entity.User'
        description: Using User as Customer for now
      customer_id:
        type: integer
      delivery_orders:
        items:
          $ref: '#/definitions/entity.DeliveryOrder'
        type: array
      discount_total:
        type: number
      grand_total:
        type: number
      id:
        type: string
      invoices:
        items:
          $ref: '#/definitions/entity.Invoice'
        type: array
      items:
        items:
          $ref: '#/definitions/entity.SalesOrderItem'
        type: array
      notes:
        type: string
      order_date:
        type: string
      order_number:
        type: string
      payment_method:
        $ref: '#/definitions/entity.PaymentMethod'
      payment_status:
        $ref: '#/definitions/entity.PaymentStatus'
      shipping_address:
        type: string
      status:
        $ref: '#/definitions/entity.SalesOrderStatus'
      sub_total:
        type: number
      tax_total:
        type: number
      updated_at:
        type: string
    type: object
...
  entity.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_login:
        type: string
      role:
        $ref: '#/definitions/entity.Role'
      role_id:
        type: integer
      status:
        $ref: '#/definitions/entity.UserStatus'
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  server.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/lugondev/erp-warehouse-simple
  description: Simple ERP Warehouse Management System API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: ERP Warehouse API
  version: "1.0"
paths:
  /api/customers:
    get:
      description: List customers with optional filtering
      parameters:
      - description: Filter by code
        in: query
        name: code
        type: string
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by type
        in: query
        name: type
        type: string
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Filter by phone number
        in: query
        name: phone_number
        type: string
      - description: Filter by loyalty tier
        in: query
        name: loyalty_tier
        type: string
      - description: Filter by city
        in: query
        name: city
        type: string
      - description: Filter by country
        in: query
        name: country
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Customer'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: List customers
      tags:
      - customers
    post:
      consumes:
      - application/json
      description: Create a new customer with the provided details
      parameters:
      - description: Customer details
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/entity.Customer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new customer
      tags:
      - customers
  /api/customers/{id}:
    delete:
      description: Delete a customer by their ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Delete a customer
      tags:
      - customers
    get:
      description: Get a customer's details by their ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Customer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Get a customer by ID
      tags:
      - customers
    put:
      consumes:
      - application/json
      description: Update a customer's details
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated customer details
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/entity.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Update a customer
      tags:
      - customers
  /api/customers/{id}/addresses:
    get:
      description: Get all addresses for a customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.CustomerAddress'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Get all addresses for a customer
      tags:
      - customers
    post:
      consumes:
      - application/json
      description: Create a new address for a customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Address details
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/entity.CustomerAddress'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.CustomerAddress'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new address for a customer
      tags:
      - customers
  /api/customers/{id}/debt:
    get:
      description: Get a customer's debt information
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CustomerDebt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Get a customer's debt information
      tags:
      - customers
    put:
      consumes:
      - application/json
      description: Update a customer's debt amount
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Debt amount
        in: body
        name: debt
        required: true
        schema:
          additionalProperties:
            type: number
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Update a customer's debt amount
      tags:
      - customers
  /api/customers/{id}/loyalty/calculate-tier:
    get:
      description: Calculate a customer's loyalty tier based on points and purchase
        history
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Calculate a customer's loyalty tier
      tags:
      - customers
  /api/customers/{id}/loyalty/points:
    put:
      consumes:
      - application/json
      description: Update a customer's loyalty points
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Loyalty points
        in: body
        name: points
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Update a customer's loyalty points
      tags:
      - customers
  /api/customers/{id}/loyalty/tier:
    put:
      consumes:
      - application/json
      description: Update a customer's loyalty tier
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Loyalty tier
        in: body
        name: tier
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Update a customer's loyalty tier
      tags:
      - customers
  /api/customers/{id}/orders:
    get:
      description: Get a summary of a customer's order history
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CustomerOrderHistory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Get a customer's order history
      tags:
      - customers
  /api/customers/addresses/{addressId}:
    delete:
      description: Delete a customer address
      parameters:
      - description: Address ID
        in: path
        name: addressId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Delete a customer address
      tags:
      - customers
    put:
      consumes:
      - application/json
      description: Update a customer address
      parameters:
      - description: Address ID
        in: path
        name: addressId
        required: true
        type: integer
      - description: Updated address details
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/entity.CustomerAddress'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CustomerAddress'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Update a customer address
      tags:
      - customers
...
  /api/v1/reports/sales/customers:
    get:
      description: Get customer sales report
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get customer sales report
      tags:
      - Reports
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"